Optimize stable branch releases to only occur on new versions
